refactor(sim): migrate parameters to config package and clean up imports
- Consolidated all parameter and related utilities from `pymob.sim.parameters` into the new `pymob.sim.config` package.
- Updated imports across the project (core, inference backends, and simulation modules) to reference the new `sim.config` location.
- Simplified `pymob/__init__.py` to expose only `Config` from the config submodule, removing outdated imports.
- Removed `pymob.sim.parameters` from the public API in `pymob/sim/__init__.py`.
- Added `pymob/sim/config/__init__.py` to re-export configuration symbols for easier access.
- This reorganization improves module structure, reduces circular dependencies, and provides a clearer, more maintainable API.
