+++ pymob/__init__.py
@@ -8,20 +8,4 @@ __version__ = "0.5.23"
 
 
 from .simulation import SimulationBase
-
-from .sim.config import (
-    Config, 
-    DataVariable, 
-)
-
-from .sim.parameters import (
-    RandomVariable,
-    Param,
-)
-
-from .solvers import (
-    SolverBase, 
-    JaxSolver
-)
-
-from .inference.base import InferenceBackend
+from .sim.config import Config
