+++ pymob/sim/__init__.py
@@ -2,6 +2,5 @@ from . import evaluator
 from . import base
 from . import solvetools
 from . import config
-from . import parameters
 from . import plot
 from . import report
