+++ pymob/sim/config/__init__.py
@@ -1,0 +1,12 @@
+from . import parameters
+from . import sections
+from . import base
+from . import casestudy_registry
+
+from .sections import *
+
+from .parameters import *
+
+from .base import *
+
+from .casestudy_registry import *
