+++ pymob/sim/config/casestudy_registry.py
